pyproject.toml
[tool.poetry]
name = "remote-comands-handler"
version = "0.1.0"
description = ""
authors = ["Stijn Bijnen <stijn.bijnen@evergen.energy>"]
readme = "README.md"
packages = [{include = "main.py"}]

[tool.poetry.dependencies]
python = "^3.11"
pymodbus = "^3.2.2"
paho-mqtt = "^1.6.1"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pylint = "^2.17.4"
pytest = "^7.3.1"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
